+++ fabricio/docker/__init__.py
@@ -2,4 +2,4 @@ from .base import Option, Attribute, BaseService
 from .image import Image, ImageNotFoundError, ImageError
 from .container import Container, ContainerNotFoundError, ContainerError
 from .registry import Registry
-from .service import Service, ServiceNotFoundError, ServiceError, Stack
+from .service import Service, ServiceNotFoundError, ServiceError, Stack, StackError